Change file structure - Added a new prop `disabledSend` to disable sent input - Block send mensagens when loading and chat empty - Added yarn hooks script - Added .mov ond vídeo files
