+++ __tests__/TestComponent.js
@@ -3,7 +3,7 @@ import { MuiThemeProvider, createMuiTheme } from '@material-ui/core/styles';
 import purple from '@material-ui/core/colors/purple';
 import green from '@material-ui/core/colors/green';
 
-import { Chat } from '../src/Chat';
+import { Chat } from '../src';
 
 // A theme with custom primary and secondary color.
 // It's optional.
