+++ package.json
@@ -80,7 +80,8 @@
     "storybook": "start-storybook -p 9001 -c .storybook",
     "storybook:build": "build-storybook -c .storybook -o docs/",
     "test": "NODE_ENV=test jest",
-    "test:windows": "set NODE_ENV=test&& jest"
+    "test:windows": "set NODE_ENV=test&& jest",
+    "hooks": "bash .git/hooks/pre-commit"
   },
   "license": "ISC",
   "husky": {
