+++ src/components/Chat/ChatStory/chatStoryFunctions.js
@@ -1,0 +1,3 @@
+export function sleep(ms) {
+  return new Promise(resolve => setTimeout(resolve, ms));
+}
