+++ src/components/Chat/Maximized/ChatHeader/ChatHeader.js
