+++ src/components/Chat/Maximized/ErrorDialog/ErrorDialog.js
