+++ src/components/Chat/Maximized/InputComposer/CustomUploader/CustomUploader.js
