+++ src/components/Chat/Maximized/InputComposer/MicRecorder/MicRecorder.js
