+++ src/components/Chat/Maximized/MessageView/Message/ApplicationLoader/ApplicationLoader.js
