+++ src/components/Chat/Maximized/MessageView/Message/ImageLoader/ImageLoader.js
