+++ src/components/Chat/Maximized/MessageView/Message/VideoLoader/VideoLoader.js
