+++ src/components/Chat/Minimized/Minimized.js
