+++ src/components/Chat/constants/CHAT_LOCATIONS.js
@@ -1,4 +1,4 @@
-export const ChatLocations = {
+export const CHAT_LOCATIONS = {
   CHAT_LIST: 0,
   MESSAGES: 1,
 };
